% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{categories}
\alias{categories}
\title{List all categories}
\usage{
categories(category_id = NULL)
}
\arguments{
\item{category_id}{A category ID.}
}
\value{
Categories as a \code{data.frame}.
}
\description{
List all categories
}
