% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{categories_graph}
\alias{categories_graph}
\title{Create graph of category hierarchy}
\usage{
categories_graph()
}
\value{
Categories as a \code{tbl_graph}.
}
\description{
Create graph of category hierarchy
}
