% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{category_products}
\alias{category_products}
\title{Products for a specific retailer}
\usage{
category_products(category_id, recursive = TRUE, ...)
}
\arguments{
\item{category_id}{A category ID.}

\item{recursive}{Should all sub-categories be included?}

\item{...}{Arguments passed through to \code{paginate()}.}
}
\value{
Details of all products for a specific category as a \code{data.frame}.
}
\description{
Products for a specific retailer
}
\examples{
# Get products for a specific category.
\dontrun{
category_id = categories() \%>\% filter(category_label == "red wine") \%>\% pull(category_id)
category_products(category_id)
}
}
