% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_response_error}
\alias{check_response_error}
\title{Check API response for errors}
\usage{
check_response_error(response)
}
\arguments{
\item{response}{}
}
\value{

}
\description{
Check API response for errors
}
