% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_response_json}
\alias{check_response_json}
\title{Check that API response is JSON}
\usage{
check_response_json(response)
}
\arguments{
\item{response}{A response object.}
}
\value{
NULL. Raises an error if the response is not JSON.
}
\description{
Check that API response is JSON
}
