% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{GET}
\alias{GET}
\title{Wrapper for httr::GET()}
\usage{
GET(url = NULL, config = list(), ...)
}
\arguments{
\item{url}{URL to retrieve.}

\item{config}{Additional configuration settings.}

\item{...}{Further named parameters.}
}
\description{
Wrapper for httr::GET()
}
