% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trundler.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieve API key}
\usage{
get_api_key()
}
\value{
API key.
}
\description{
Retrieve API key
}
\examples{
set_api_key("8f9f3c4e-5dd6-4bff-3a2c-592b45cf2437")
get_api_key()
}
