% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{palette}
\alias{palette}
\alias{TRUNDLER_SLATE}
\alias{TRUNDLER_ORANGE}
\alias{TRUNDLER_BLUE}
\alias{TRUNDLER_GREY}
\alias{TRUNDLER_GREEN}
\alias{TRUNDLER_RED}
\title{Trundler palette}
\format{











}
\usage{
TRUNDLER_SLATE

TRUNDLER_ORANGE

TRUNDLER_BLUE

TRUNDLER_GREY

TRUNDLER_GREEN

TRUNDLER_RED
}
\description{
A selection of "standard" colours to be used in Trundler plots.
}
\keyword{datasets}
