% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{param_boolean}
\alias{param_boolean}
\title{Convert URL parameter into Boolean.}
\usage{
param_boolean(bool)
}
\arguments{
\item{bool}{A Boolean string (either "true" or "false").}
}
\value{
A Boolean value (either TRUE or FALSE).
}
\description{
Convert URL parameter into Boolean.
}
