% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_complete.R
\name{price_complete}
\alias{price_complete}
\title{Interpolate prices for a specific product within a given date range.}
\usage{
price_complete(product_id, date_min = NULL, date_max = NULL)
}
\arguments{
\item{product_id}{A product ID.}

\item{date_min}{The minimum date for the interpolation range as a character string in the format "YYYY-mm-dd".  Defaults to the minimum date available for the specified product ID.}

\item{date_max}{The maximum date for the interpolation range as a character string in the format "YYYY-mm-dd".  Defaults to the maximum date available for the specified product ID.}
}
\value{
Interpolated prices for the specified (or default) date range as a \code{data.frame}.
}
\description{
Interpolate prices for a specific product within a given date range.
}
\examples{
# Interpolate prices for a specific product within a given date range.
\dontrun{
price_complete(product_id = 1, date_min = "2020-01-01", date_max = "2020-12-31")
}
}
