% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\name{product}
\alias{product}
\title{Details for a specific product}
\usage{
product(product_id)
}
\arguments{
\item{product_id}{A product ID.}
}
\value{
Product details as a \code{data.frame}.
}
\description{
Details for a specific product
}
\examples{
# Details of a specific product.
\dontrun{
product(1)
}
}
