% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\name{products}
\alias{products}
\title{Find products by name or brand}
\usage{
products(name = NA, brand = NA, ...)
}
\arguments{
\item{name}{Filter by product name (treated as a regular expression).}

\item{brand}{Filter by product brand (treated as a regular expression).}

\item{...}{Arguments passed through to \code{paginate()}.}
}
\value{
Product details as a \code{data.frame}.
}
\description{
Find products by name or brand
}
\examples{
\dontrun{
products(name = "coffee")
products(brand = "Illy")
products(name = "coffee", brand = "Illy")
}
}
