% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retailer.R
\name{retailer}
\alias{retailer}
\title{Details of a retailer}
\usage{
retailer(retailer_id = NA)
}
\arguments{
\item{retailer_id}{A retailer ID.}
}
\value{
Details of either all retailers or a specific retailer as a \code{data.frame}.
}
\description{
Details of a retailer
}
\examples{
# Get a table of all retailers.
\dontrun{
retailer()
}
# Get details of specific retailer.
\dontrun{
retailer(1)
}
}
