% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retail.R
\name{set_api_server}
\alias{set_api_server}
\title{Set IP and port for API server}
\usage{
set_api_server(ip, port)
}
\arguments{
\item{ip}{IP address}

\item{port}{Port}
}
\value{

}
\description{
Set IP and port for API server
}
\examples{
# Use local server
set_api_server("0.0.0.0", 8080)
}
