% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trundler.R
\name{set_rapidapi_key}
\alias{set_rapidapi_key}
\title{Set RapidAPI key}
\usage{
set_rapidapi_key(api_key)
}
\arguments{
\item{api_key}{API key for RapidAPI.}
}
\description{
Set RapidAPI key
}
\examples{
set_rapidapi_key("5a1ae0ce24mshd483dae6ab7308dp129ef6jsn1f473053d6b0")
}
