% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_trundler_title}
\alias{theme_trundler_title}
\title{Format plot title, subtitle and caption}
\usage{
theme_trundler_title()
}
\description{
This is a separate function so that it can be applied to plots which are
constructed using {patchwork}.
}
