% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_trundler_watermark}
\alias{theme_trundler_watermark}
\title{Add watermark to plot}
\usage{
theme_trundler_watermark(label = "WIP", fontsize = 256)
}
\arguments{
\item{label}{}
}
\description{
Make this the first layer in the plot so that it's below all the content.
}
