% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{categories_graph}
\alias{categories_graph}
\title{Create graph of category hierarchy}
\usage{
categories_graph(taxonomy = NULL)
}
\arguments{
\item{taxonomy}{A data frame returned by \code{} or NULL.}
}
\value{
Categories as a \code{tbl_graph}.
}
\description{
Create graph of category hierarchy
}
\examples{
\dontrun{
library(ggraph)

graph <- categories_graph()
# In principle there should be graph data available, but let's check.
if (!is.null(graph)) {
  ggraph(graph, 'tree') +
    geom_edge_link() +
    geom_node_point() +
    geom_node_label(aes(label = category_label)) +
    theme_graph()
}
}
}
