% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_response_error}
\alias{check_response_error}
\title{Check API response for errors}
\usage{
check_response_error(response)
}
\arguments{
\item{response}{A response object.}
}
\value{
NULL. Raises an error if response has an error code.
}
\description{
Check API response for errors
}
