% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_response_json}
\alias{check_response_json}
\title{Check that API response is JSON}
\usage{
check_response_json(response)
}
\arguments{
\item{response}{}
}
\value{

}
\description{
Check that API response is JSON
}
