% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\name{product_prices}
\alias{product_prices}
\title{Price history for a specific product}
\usage{
product_prices(product_id, ...)
}
\arguments{
\item{product_id}{A product ID.}

\item{...}{Arguments passed through to \code{paginate()}.}
}
\value{
Price history as a \code{data.frame}.
}
\description{
Price history for a specific product
}
\examples{
# Price history for a specific product.
\dontrun{
product_prices(1)
}
}
