% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retailer.R
\name{retailer_products}
\alias{retailer_products}
\title{Products for a specific retailer}
\usage{
retailer_products(retailer_id, name = NA, ...)
}
\arguments{
\item{retailer_id}{A retailer ID.}

\item{name}{A product name to search for (treated as a regular expression).}

\item{...}{Arguments passed through to \code{paginate()}.}
}
\value{
Details of all products for a specific retailer as a \code{data.frame}.
}
\description{
Products for a specific retailer
}
\examples{
# Get products for a specific retailer.
\dontrun{
retailer_products(1)
retailer_products(9, name = "Nescafe")
}
}
