% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trundler.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{API key.}
}
\description{
Set API key
}
\examples{
set_api_key("8f9f3c4e-5dd6-4bff-3a2c-592b45cf2437")
}
