% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_trundler_title}
\alias{theme_trundler_title}
\title{Format plot title, subtitle and caption}
\usage{
theme_trundler_title(subtitle = TRUE)
}
\arguments{
\item{subtitle}{Set to TRUE by default.  If FALSE, then increase the margin size to prevent collisions between the title and the plot.}
}
\description{
This is a separate function so that it can be applied to plots which are
constructed using {patchwork}.
}
