% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_trundler_title}
\alias{theme_trundler_title}
\title{Format plot title, subtitle and caption}
\usage{
theme_trundler_title(title = TRUE, subtitle = TRUE, caption = TRUE)
}
\arguments{
\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}
}
\description{
Format plot title, subtitle and caption
}
